+++ src/features/treeViewer/hooks/useNodesRuntimeUpdates.ts
@@ -335,6 +335,17 @@ export const useNodesRuntimeUpdates = (playbackInterval: number) => {
 
         case "setRightChild":
           if (frame.prevArgs) {
+            if ("revertChildId" in slice.actions) {
+              dispatch(
+                slice.actions.revertChildId({
+                  name: treeName,
+                  data: {
+                    id: frame.nodeId,
+                    ...frame.args,
+                  },
+                }),
+              );
+            }
             applyFrame({
               ...frame,
               name: "setRightChild",